+++ matlab-extension/frequency/frequency-acceleration/src/main/java/com/tp/matlab/extension/frequency/acceleration/TotalValue.java
@@ -15,12 +15,11 @@ import static java.util.stream.Collectors.toList;
  */
 @RequiredArgsConstructor
 class TotalValue {
-
     /**
      * 源数据
      */
     private final List<Double> a;
-    private final Long fs;
+    private final Integer fs;
     private final Integer fmin;
     private final Integer fmax;
     private final Integer m;
