+++ matlab-extension/frequency/frequency-envolope/src/main/java/com/tp/matlab/extension/frequency/envolope/HannFilt.java
@@ -13,11 +13,10 @@ import java.util.List;
  */
 @RequiredArgsConstructor
 class HannFilt {
-
     /**
      * 采样频率
      */
-    private final Long fs;
+    private final Integer fs;
     /**
      * 源数据
      */
