+++ matlab-extension/time/time-velocity/time-velocity-common/src/main/java/com/tp/matlab/extension/time/velocity/common/A2v.java
@@ -20,7 +20,6 @@ import static java.util.stream.Collectors.toList;
  */
 @RequiredArgsConstructor
 public class A2v {
-
     /**
      * 源数据
      */
@@ -28,7 +27,7 @@ public class A2v {
     /**
      * 采样频率
      */
-    private final Long fs;
+    private final Integer fs;
     /**
      * 低频截止
      */
