+++ matlab-extension/time/time-velocity/time-velocity-common/src/main/java/com/tp/matlab/extension/time/velocity/common/Filt.java
@@ -19,7 +19,6 @@ import static java.util.stream.Collectors.toList;
  */
 @RequiredArgsConstructor
 public class Filt {
-
     /**
      * 源数据
      */
@@ -27,7 +26,7 @@ public class Filt {
     /**
      * 采样频率
      */
-    private final Long fs;
+    private final Integer fs;
     /**
      * 低频截止
      */
